3/11/2005 - giwo (0.98-2.8d) 	Included jr's fixes for cygwin compatability. 	Began work on allowing UOX to compile under MinGW. 	Many minor gcc warning fixes.

+++ source/skills.cpp
@@ -3733,7 +3733,7 @@ void cSkills::AdvanceStats( CChar *s, UI08 sk, bool skillsuccess )
 	ALSO this prevents char's stats from FREEZEING when they reach the cap*/
 	
 	//k will only take away from other stats if within 10 of the server stat cap
-	if( static_cast<int>(ttlStats) >= RandomNum( ServStatCap-10, ServStatCap ) )
+	if( ttlStats >= RandomNum( ServStatCap-10, ServStatCap ) )
 	{
 		UI08 skillDrop = 0;
